refactored Main and NavBar files.
